+++ race_manager.go
@@ -72,22 +72,12 @@ func (rm *RaceManager) CurrentRace() (*ServerConfig, EntryList) {
 
 var ErrEntryListTooBig = errors.New("servermanager: EntryList exceeds MaxClients setting")
 
-type RaceEvent interface {
-	IsChampionship() bool
-	IsRaceWeekend() bool
-	OverrideServerPassword() bool
-	ReplacementServerPassword() string
-	EventName() string
-	EventDescription() string
-	GetURL() string
-}
-
-func (rm *RaceManager) applyConfigAndStart(raceConfig CurrentRaceConfig, entryList EntryList, loop bool, event RaceEvent) error {
+func (rm *RaceManager) applyConfigAndStart(event RaceEvent) error {
 	rm.mutex.Lock()
 	defer rm.mutex.Unlock()
 
 	// Reset the stored session types if this isn't a looped race
-	if !loop {
+	if !event.IsLooping() {
 		rm.clearLoopedRaceSessionTypes()
 	}
 
@@ -97,6 +87,15 @@ func (rm *RaceManager) applyConfigAndStart(raceConfig CurrentRaceConfig, entryLi
 	if err != nil {
 		return err
 	}
+
+	if serverOpts.RestartEventOnServerManagerLaunch == 1 {
+		if err := rm.store.ClearLastRaceEvent(); err != nil {
+			logrus.WithError(err).Errorf("Could not clear last race event")
+		}
+	}
+
+	raceConfig := event.GetRaceConfig()
+	entryList := event.GetEntryList()
 
 	config := ServerConfig{
 		CurrentRaceConfig:  raceConfig,
@@ -207,7 +206,7 @@ func (rm *RaceManager) applyConfigAndStart(raceConfig CurrentRaceConfig, entryLi
 		return err
 	}
 
-	if !loop {
+	if !event.IsLooping() {
 		_ = rm.notificationManager.SendRaceStartMessage(config, event)
 	}
 
@@ -346,8 +345,9 @@ func (rm *RaceManager) SetupQuickRace(r *http.Request) error {
 
 	quickRace.MaxClients = numPitboxes
 
-	return rm.applyConfigAndStart(quickRace, entryList, false, &QuickRace{
+	return rm.applyConfigAndStart(&QuickRace{
 		RaceConfig: quickRace,
+		EntryList:  entryList,
 	})
 }
 
@@ -672,7 +672,7 @@ func (rm *RaceManager) SetupCustomRace(r *http.Request) error {
 			return nil
 		}
 
-		return rm.applyConfigAndStart(completeConfig.CurrentRaceConfig, entryList, false, race)
+		return rm.applyConfigAndStart(race)
 	}
 }
 
@@ -1002,7 +1002,9 @@ func (rm *RaceManager) StartCustomRace(uuid string, forceRestart bool) error {
 		race.RaceConfig.LoopMode = 1
 	}
 
-	return rm.applyConfigAndStart(race.RaceConfig, race.EntryList, forceRestart, race)
+	race.Loop = forceRestart
+
+	return rm.applyConfigAndStart(race)
 }
 
 func (rm *RaceManager) ScheduleRace(uuid string, date time.Time, action string, recurrence string) error {
@@ -1095,7 +1097,15 @@ func (rm *RaceManager) StartScheduledRace(race *CustomRace) error {
 
 func (rm *RaceManager) ScheduleNextFromRecurrence(race *CustomRace) error {
 	// set the scheduled time to the next iteration of the recurrence rule
-	return rm.ScheduleRace(race.UUID.String(), rm.FindNextRecurrence(race, race.Scheduled), "add", "already-set")
+	nextRecurrence := rm.FindNextRecurrence(race, race.Scheduled)
+
+	if nextRecurrence.IsZero() {
+		// no recurrence was found (likely the recurrence had an UNTIL date)
+		race.ClearRecurrenceRule()
+		return rm.store.UpsertCustomRace(race)
+	}
+
+	return rm.ScheduleRace(race.UUID.String(), nextRecurrence, "add", "already-set")
 }
 
 func (rm *RaceManager) FindNextRecurrence(race *CustomRace, start time.Time) time.Time {
@@ -1110,6 +1120,8 @@ func (rm *RaceManager) FindNextRecurrence(race *CustomRace, start time.Time) tim
 
 	if next.After(time.Now()) {
 		return next
+	} else if next.IsZero() {
+		return next
 	} else {
 		return rm.FindNextRecurrence(race, next)
 	}
@@ -1120,6 +1132,14 @@ func (rm *RaceManager) DeleteCustomRace(uuid string) error {
 
 	if err != nil {
 		return err
+	}
+
+	if !race.Scheduled.IsZero() {
+		err := rm.ScheduleRace(uuid, time.Time{}, "remove", "")
+
+		if err != nil {
+			return err
+		}
 	}
 
 	return rm.store.DeleteCustomRace(race)